authのディレクトリ構造を変更
